+++ benches/bench.rs
@@ -29,6 +29,26 @@ fn bench_fasterpam(b: &mut Bencher) {
 }
 
 #[bench]
+fn bench_rand_fasterpam(b: &mut Bencher) {
+	let n = 100;
+	let mut rng = StdRng::seed_from_u64(42);
+	let mut mat = Array2::<i32>::from_elem((n, n), 0);
+	for i in 0..n {
+		for j in (i + 1)..n {
+			let v = rng.gen_range(1..100);
+			mat[[i, j]] = v;
+			mat[[j, i]] = v;
+		}
+	}
+	b.iter(|| {
+		let mut med = vec![0, 1, 2, 3, 4];
+		let (loss, assignment, _, _): (i32, _, _, _) = rand_fasterpam(&mat, &mut med, 100, &mut rng);
+		black_box(loss);
+		black_box(assignment);
+	});
+}
+
+#[bench]
 fn bench_fastpam1(b: &mut Bencher) {
 	let n = 100;
 	let mut rng = StdRng::seed_from_u64(42);