+++ src/kmedoids.rs
@@ -318,11 +318,15 @@ where
 /// ```
 #[cfg(feature = "rand")]
 #[inline]
-pub fn random_initialization(n: usize, k: usize, mut rng: &mut impl rand::Rng) -> Vec<usize> {
-	rand::seq::index::sample(&mut rng, n, k).into_vec()
+pub fn random_initialization(n: usize, k: usize, rng: &mut impl rand::Rng) -> Vec<usize> {
+	rand::seq::index::sample(rng, n, k).into_vec()
 }
 
 /// Run the FasterPAM algorithm.
+/// 
+/// If used multiple times, it is better to additionally shuffle the input data,
+/// to increase randomness of the solutions found and hence increase the chance
+/// of finding a better solution.
 ///
 /// * type `M` - matrix data type such as `ndarray::Array2` or `kmedoids::arrayadapter::LowerTriangle`
 /// * type `N` - number data type such as `u32` or `f64`
@@ -383,6 +387,105 @@ where
 		// println!("Iteration {} before {}", iter, loss);
 		let swaps_before = n_swaps;
 		for j in 0..n {
+			if j == lastswap {
+				break;
+			}
+			if j == data[j].near.i as usize {
+				continue; // This already is a medoid
+			}
+			let (change, b) = find_best_swap(mat, &removal_loss, &data, j);
+			if change >= L::zero() {
+				continue; // No improvement
+			}
+			n_swaps += 1;
+			lastswap = j;
+			// perform the swap
+			let newloss = do_swap(mat, med, &mut data, b, j);
+			// println!("{} + {} = {} vs. {}", loss, change, loss + change, newloss);
+			if newloss >= loss {
+				break; // Probably numerically unstable now.
+			}
+			loss = newloss;
+			update_removal_loss(&data, &mut removal_loss);
+		}
+		if n_swaps == swaps_before {
+			break; // converged
+		}
+	}
+	// println!("final loss: {}", loss);
+	// println!("number of swaps: {}", n_swaps);
+	let assi = data.iter().map(|x| x.near.i as usize).collect();
+	(loss, assi, iter, n_swaps)
+}
+
+/// Run the FasterPAM algorithm with additional randomization.
+/// 
+/// This increases the chance of finding a better solution when used multiple times,
+/// as it decreases the dependency on the input data order.
+///
+/// * type `M` - matrix data type such as `ndarray::Array2` or `kmedoids::arrayadapter::LowerTriangle`
+/// * type `N` - number data type such as `u32` or `f64`
+/// * type `L` - number data type such as `i64` or `f64` for the loss (must be signed)
+/// * `mat` - a pairwise distance matrix
+/// * `med` - the list of medoids
+/// * `maxiter` - the maximum number of iterations allowed
+/// * `rng` - random number generator for shuffling the input data
+///
+/// returns a tuple containing:
+/// * the final loss
+/// * the final cluster assignment
+/// * the number of iterations needed
+/// * the number of swaps performed
+///
+/// ## Panics
+///
+/// * panics when the dissimilarity matrix is not square
+/// * panics when k is 0 or larger than N
+///
+/// ## Example
+/// Given a dissimilarity matrix of size 4 x 4, use:
+/// ```
+/// let data = ndarray::arr2(&[[0,1,2,3],[1,0,4,5],[2,4,0,6],[3,5,6,0]]);
+/// let mut meds = kmedoids::random_initialization(4, 2, &mut rand::thread_rng());
+/// let (loss, assi, n_iter, n_swap): (f64, _, _, _) = kmedoids::rand_fasterpam(&data, &mut meds, 100, &mut rand::thread_rng());
+/// println!("Loss is: {}", loss);
+/// ```
+#[cfg(feature = "rand")]
+pub fn rand_fasterpam<M, N, L>(
+	mat: &M,
+	med: &mut Vec<usize>,
+	maxiter: usize,
+	rng: &mut impl rand::Rng,
+) -> (L, Vec<usize>, usize, usize)
+where
+	N: Zero + PartialOrd + Copy + std::fmt::Display,
+	L: AddAssign + Signed + Zero + PartialOrd + Copy + From<N> + std::fmt::Display,
+	M: ArrayAdapter<N>,
+{
+	let n = mat.len();
+	let k = med.len();
+	if k == 1 {
+		let assi = vec![0; n];
+		let (swapped, loss) = choose_medoid_within_partition::<M, N, L>(mat, &assi, med, 0);
+		return (loss, assi, 1, if swapped { 1 } else { 0 });
+	}
+	let mut data = Vec::<Rec<N>>::with_capacity(n);
+	let mut loss = initial_assignment(mat, med, &mut data);
+	#[cfg(feature = "assertions")]
+	debug_validate_assignment(mat, med, &data);
+
+	// println!("Initial loss is {}", loss);
+	let mut removal_loss = vec![L::zero(); k];
+	update_removal_loss(&data, &mut removal_loss);
+	let mut lastswap = n;
+	let mut n_swaps = 0;
+	let mut iter = 0;
+	let seq = rand::seq::index::sample(rng, n, n); // random shuffling
+	while iter < maxiter {
+		iter += 1;
+		// println!("Iteration {} before {}", iter, loss);
+		let swaps_before = n_swaps;
+		for j in seq.iter() {
 			if j == lastswap {
 				break;
 			}
@@ -1065,6 +1168,29 @@ mod tests {
 		assert_array(assi, vec![0, 0, 0, 0, 0], "assignment not as expected");
 		assert_array(meds, vec![0], "medoids not as expected");
 		assert_eq!(sil, 0., "Silhouette not as expected");
+	}
+
+	#[cfg(feature = "rand")]
+	use rand::{rngs::StdRng, SeedableRng};
+	#[cfg(feature = "rand")]
+	use crate::rand_fasterpam;
+	#[cfg(feature = "rand")]
+	#[test]
+	fn testrand_fasterpam() {
+		let data = LowerTriangle {
+			n: 5,
+			data: vec![1, 2, 3, 4, 5, 6, 7, 8, 9, 1],
+		};
+		let mut meds = vec![0, 1];
+		let mut rng = StdRng::seed_from_u64(1);
+		let (loss, assi, n_iter, n_swap): (i64, _, _, _) = rand_fasterpam(&data, &mut meds, 10, &mut rng);
+		let (sil, _): (f64, _) = silhouette(&data, &assi, false);
+		assert_eq!(loss, 4, "loss not as expected");
+		assert_eq!(n_swap, 1, "swaps not as expected");
+		assert_eq!(n_iter, 2, "iterations not as expected");
+		assert_array(assi, vec![0, 0, 0, 1, 1], "assignment not as expected");
+		assert_array(meds, vec![0, 4], "medoids not as expected");
+		assert_eq!(sil, 0.7522494172494172, "Silhouette not as expected");
 	}
 
 	#[test]